Use features to seperate the macro and build plugin components
