+++ test/Cargo.toml
@@ -5,9 +5,7 @@ authors = ["Michael Layzell <michael@thelayzells.com>"]
 build = "build.rs"
 
 [build-dependencies]
-cpp = { path = "../" }
+cpp = { path = "../", features = ["build"] }
 
 [dependencies]
-cpp = { path = "../" }
-
-
+cpp = { path = "../", features = ["macro"] }
